ECR-3779 Review fixes
